requirements.txt
azure-cognitiveservices-speech
Flask
Flask-Cors
ffmpeg-python
